[#134] clarified structure of _plot_as_array.py
